requirements.txt
streamlit
pandas
openpyxl
rapidfuzz
python-docx
PyMuPDF